+++ custom_components/local_weather_forecast/sensor.py
@@ -184,7 +184,7 @@ class LocalWeatherForecastEntity(RestoreEntity, SensorEntity):
             "name": "Local Weather Forecast",
             "manufacturer": "Local Weather Forecast",
             "model": "Zambretti Forecaster",
-            "sw_version": "3.1.25",
+            "sw_version": "3.1.26",
         }
 
     async def _wait_for_entity(